refactor: move ticker to eventsource package
This commit extracts the ticker event source into a dedicated package so
that it can be used for multiple controllers.
